separate datagen models from src code
